Lint: Integral division result
Integral division result cast to double or float
This code casts the result of an integral division (e.g., int or long division) operation to double or float. Doing division on integers truncates the result to the integer value closest to zero. The fact that the result was cast to double suggests that this precision should have been retained. What was probably meant was to cast one or both of the operands to double before performing the division. Here is an example:
int x = 2;
int y = 5;
// Wrong: yields result 0.0
double value1 = x / y;

// Right: yields result 0.4
double value2 = x / (double) y;

Bug kind and pattern: ICAST - ICAST_IDIV_CAST_TO_DOUBLE

+++ src/main/java/handWavey/HandsState.java
@@ -333,7 +333,7 @@ public class HandsState {
         double offsetRoll = handedRoll + this.pi + (segmentWidth/2) - userOffset;
 
         // Work out which segment we are in, then rotate it back because we rotated to make the entire range positive.
-        int segmentNumber = (int) Math.floor((offsetRoll / segmentWidth) - (segmentCount / 2));
+        int segmentNumber = (int) Math.floor(((double)offsetRoll / (double)segmentWidth) - ((double)segmentCount / 2));
         while (segmentNumber < 0) {
             segmentNumber += segmentCount;
         }
